Abstract entity class
